Impulse sets velocity for a pose. Change units of Movable.evolve() from milliseconds to seconds.

+++ cosmos/src/com/aimlessblade/cosmos/geo/Movable.java
@@ -1,7 +1,7 @@
 package com.aimlessblade.cosmos.geo;
 
 public interface Movable {
-    void evolve(long dt);
+    void evolve(double dt);
 
     void impulse(double vx, double vy, double vz);
